Feat: Correcciones en la base de datos e implementaciones en los componentes de estadosPagos y movimientos

+++ database/migrations/2023_07_19_211143_create_colaboradors_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('colaboradors', function (Blueprint $table) {
             $table->integer('idcolaborador')->autoIncrement()->comment('Llave primaria');
             $table->string('nombrecompleto', 50)->comment('Nombre Completo');
-            $table->integer('telefono')->comment('Numero Telefono');
+            $table->bigInteger('telefono')->comment('Numero Telefono');
             $table->string('correo', 50)->comment('Correo Electronico');
             $table->date('fcnacimiento')->comment('Fecha de nacimiento');
             $table->timestamps();
